ref: rename dataValidator to kebab-case
