+++ src/entities/auth-form/ui/form/form.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { authValidator } from '../../../dataValidator'
+import { authValidator } from '../../../data-validator'
 import { AuthForm } from './form'
 
 const meta = {
