+++ src/entities/data-validator/auth.ts
