+++ src/entities/data-validator/index.ts
